Renamed AccountDetails and made minor modifications
Renamed the `AccountDetails.java` file and moved it to `ua.com.javarush.oleksandr.reddit.redditcloneabstract.security` package. Also, a validation constraint has been added to the User object, and username has been replaced with email to get the username. These changes aim to align with the user authentication standards of the application.

Signed-off-by: Aleksei Os. <oal.aleos@gmail.com>