chore: add initialization script for Reddit data
This commit includes an SQL script for initializing a set of sample data on Reddit. The script truncates existing data and resets identity sequences for the "user", "subreddit", "post", and "comment" tables. Then, it populates these tables with sample data representative of real-world usage. The aim is to provide a consistent test data set for the application.

Signed-off-by: Aleksei Os. <oal.aleos@gmail.com>